# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Expert Quotes Requests Service",
    "description": "Bronze wrapper for journalist quote-request providers (Featured.com first; HARO/SOS/Qwoted later). Owns Featured.com auth lifecycle, rate-limit budgeting, cursor / sync state, and append-only bronze raw payloads.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3055"
    }
  ],
  "components": {
    "schemas": {
...
      "OpportunitiesListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeaturedOpportunityRow"
            }
          },
          "nextSince": {
            "type": "string",
            "nullable": true
          },
          "refreshed": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "nextSince",
          "refreshed"
        ]
      },
      "FeaturedOpportunityRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "externalId": {
            "type": "string"
          },
          "featuredQuestionId": {
            "type": "integer",
            "nullable": true
          },
          "opportunityText": {
            "type": "string"
          },
          "mediaOutlet": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "pitchUrl": {
            "type": "string",
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "nullable": true
          },
          "raw": {
            "nullable": true
          },
          "firstSeenAt": {
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "externalId",
          "featuredQuestionId",
          "opportunityText",
          "mediaOutlet",
          "source",
          "pitchUrl",
          "deadline",
          "firstSeenAt",
          "lastSeenAt"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "OpportunityRefreshResponse": {
        "type": "object",
        "properties": {
          "refreshed": {
            "type": "boolean"
          },
          "inserted": {
            "type": "integer"
          },
          "updated": {
            "type": "integer"
          },
          "skipped": {
            "type": "integer"
          }
        },
        "required": [
          "refreshed",
          "inserted",
          "updated",
          "skipped"
        ]
      },
      "SubmissionStatusResponse": {
        "type": "object",
        "properties": {
          "statuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubmissionStatusEntry"
            }
          }
        },
        "required": [
          "statuses"
        ]
      },
      "SubmissionStatusEntry": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string"
          },
          "submitted": {
            "type": "boolean"
          },
          "lastStatus": {
            "type": "string",
            "nullable": true
          },
          "submittedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "externalId",
          "submitted",
          "lastStatus",
          "submittedAt"
        ]
      },
      "SubmissionStatusRequest": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string",
            "format": "uuid"
          },
          "externalIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "maxItems": 500
          }
        },
        "required": [
          "brandId",
          "externalIds"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/orgs/featured/opportunities": {
      "get": {
        "summary": "List bronze Featured opportunities (lazy-refreshes from Featured)",
        "description": "Lazy-refreshes from Featured if last refresh > FEATURED_OPPORTUNITY_TTL_MS. Supports cursor pagination via `?since=<ISO>` filtered against `first_seen_at`. Response carries `nextSince` for chaining.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "since",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 500
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "When set, switches to per-brand delivery mode: returns only opportunities never delivered to this (org, brand) and records them as delivered. Consecutive calls return disjoint sets; exhausted → empty array. `nextSince` is null in this mode (the delivery ledger is the cursor). When absent, the legacy org-scoped `since` timestamp cursor is used."
            },
            "required": false,
            "description": "When set, switches to per-brand delivery mode: returns only opportunities never delivered to this (org, brand) and records them as delivered. Consecutive calls return disjoint sets; exhausted → empty array. `nextSince` is null in this mode (the delivery ledger is the cursor). When absent, the legacy org-scoped `since` timestamp cursor is used.",
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Opportunities + cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunitiesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Featured.com unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/featured/opportunities/refresh": {
      "post": {
        "summary": "Force-refresh opportunities from Featured (bypass TTL)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Refresh stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityRefreshResponse"
                }
              }
            }
          },
          "502": {
            "description": "Featured.com unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/featured/opportunities/submission-status": {
      "post": {
        "summary": "Authoritative submitted-status for a set of opportunities, per (org, brand)",
        "description": "Returns, for each requested `externalId`, whether it has been submitted-successfully to Featured for this (org, brand). `submitted` is true only when a ledger row with status `submitted` exists; `error`/pending/absent → `submitted: false` (still offerable, AC4). Keyed on the atomic single `brandId` + the same `externalId` the opportunities feed exposes.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Per-opportunity submitted-status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}